+++ src/@types/resources.ts
@@ -1,0 +1,9 @@
+import en from '@i18n/en.json';
+import es from '@i18n/es.json';
+
+const resources = {
+  en,
+  es,
+} as const;
+
+export default resources;