+++ src/utils/constants/api.constants.ts
@@ -1,0 +1,6 @@
+import { ApiVersion } from '../enums/api.enums';
+
+export const USER_SERVICE: string = 'users';
+export const POSTS_SERVICE: string = 'posts';
+
+export const CURR_API_VERSION: ApiVersion = ApiVersion.V1;