+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This document lists breaking changes for each major release.
 
+## v0.0.3 (20/11/2024)
+
+Feature: KAN-2, KAN-10, KAN-16, KAN-26, KAN-27, KAN-28, Update project structure and add login, signup, logout.
+
+- Configure imports, include types and tsx.
+- Create constants, enums, errors, validations.
+- Add fonts, package.
+- Add providers, routes and Snackbar to notifications.
+- Add reset and normalize css.
+- Create store, slice user, state and action with zustand.
+- Create services with axios.
+- Create routes, public, private and protected routes.
+- Create initial components to NotFound, Home, SignUp, Login, Navbar, Loading.
+- Create models to typscript interfaces, types.
+- Create hooks to mutation, queries, forms, alert notifications, navbar.
+- Create component to alert notifications with Notistack.
+- Add components to SignUp, Login, Navbar, Loading.
+- Add image to signup, login and logo.
+- Create adapter to user.
+- Create i18n files to translations and validations translations.
+- Generate initial config to tests and first e2e test to SignUp.
+- Add types to i18next, react scripts, resources i18n.
+
 ## v0.0.2 (17/09/2024)
 
 Feature: KAN-12, Modify structure and update config.